+++ core/src/raw/adapters/kv/backend.rs
@@ -74,6 +74,7 @@ impl<S: Adapter> Accessor for Backend<S> {
         if cap.read {
             cap.read_can_seek = true;
             cap.read_can_next = true;
+            cap.read_with_range = true;
             cap.stat = true;
         }