+++ core/src/services/gcs/writer.rs
@@ -58,7 +58,6 @@ impl GcsWriter {
             &percent_encode_path(&self.path),
             Some(bs.len()),
             self.op.content_type(),
-            self.op.cache_control(),
             AsyncBody::Bytes(bs),
         )?;