+++ core/src/services/hdfs/backend.rs
@@ -93,6 +93,11 @@ use crate::*;
 /// ```
 ///
 /// `CLASSPATH` is not set correctly or your hadoop installation is incorrect.
+///  
+/// To set `CLASSPATH`:
+/// ```shell
+/// export CLASSPATH=$(find $HADOOP_HOME -iname "*.jar" | xargs echo | tr ' ' ':'):${CLASSPATH}
+/// ```
 ///
 /// # Example
 ///
@@ -235,9 +240,13 @@ impl Accessor for HdfsBackend {
             .set_capability(Capability {
                 read: true,
                 read_can_seek: true,
+                read_with_range: true,
+
                 write: true,
                 list: true,
                 blocking: true,
+
+                list_with_delimiter_slash: true,
                 ..Default::default()
             });