+++ core/tests/behavior/blocking_list.rs
@@ -49,7 +49,6 @@ macro_rules! behavior_blocking_list_test {
                             Ok(())
                         },
                         None => {
-                            log::warn!("service {} not initiated, ignored", opendal::Scheme::$service);
                             Ok(())
                         }
                     }